requirements.txt
torch>=2.0.0
transformers>=4.40.0
Pillow>=10.0.0
requests>=2.31.0
python-dotenv>=1.0.0
loguru>=0.7.0
httpx>=0.27.0
tenacity>=8.2.0